feat:
Exercise

In the lesson example, there is a problem with the code used to display data in the console. It consists in the fact that we have to repeat it for each client. What's more, if we want to change the way the data is displayed, i.e. add some formatting, we are forced to introduce corrections in several places in the code.

Make such changes to the code to hide the logic involved in presenting the results to the user in a separate class and method. Also add the functionality that if the firstName or lastName field is not set in the client, the greeting will be displayed in one of the following form:

"Hello Jan" - when the lastName field is null,
"Hello Jan Kowalski" - when both firstName and lastName are different from null,
"Hello, Mr. / Mrs. Kowalski" - if the firstName field is null,
"Hello Stranger" - if both firstName and lastName are nulls.
If the firstName or lastName field is not set, the client object is created, e.g. like this:

Client client1 = new Client (null, "Kowalski", true);